+++ Problem1.java
@@ -1,0 +1,5 @@
+public class Problem1 {
+    public static void main(String[] args) {
+        System.out.println("Hello");
+    }
+}
